// src/interfaces/cards.interface.ts
export interface Cards {
  id: string;
  lot: string;
  rank: string;
  imgSrc: string;
  images: string[];
  mark: string;
  model: string;
  engine_capacity: number;
  year_of_creation: number;
  equipment: string;
  transmission: string;
  mileage: number;
  sale: boolean;
  body: string;
  alltime: string;
  dateAuc: string;
}